% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypsography.R
\name{generate_hypsography}
\alias{generate_hypsography}
\title{generate_hypsography}
\usage{
generate_hypsography(nlayers, max_depth, surface_area, layer_thickness = NA)
}
\arguments{
\item{nlayers}{integer number of layers}

\item{max_depth}{numeric max depth}

\item{surface_area}{numeric surface area}

\item{layer_thickness}{numeric optional overrides nlayers#'}
}
\description{
generate_hypsography
}
\examples{
nlayers <- 14
max_depth <- 18.288000
surface_area <- 637641.6
res <- generate_hypsography(nlayers, max_depth, surface_area)
plot(res[,c("area", "depth_linear")],
    ylim = c(max(res$depth_linear), min(res$depth_linear)))
points(res[,c("area", "depth_ellipsoid")], col = "red", pch = 19)

}
